+++ packages/web3-eth/test/integration/defaults.test.ts
@@ -468,7 +468,8 @@ describe('defaults', () => {
 			});
 			expect(eth2.blockHeaderTimeout).toBe(4);
 		});
-		it('enableExperimentalFeatures', () => {
+
+		it('enableExperimentalFeatures useSubscriptionWhenCheckingBlockTimeout', () => {
 			// default
 			expect(web3Eth.enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout).toBe(
 				false,
@@ -476,7 +477,10 @@ describe('defaults', () => {
 
 			// after set
 			web3Eth.setConfig({
-				enableExperimentalFeatures: { useSubscriptionWhenCheckingBlockTimeout: true },
+				enableExperimentalFeatures: {
+					useSubscriptionWhenCheckingBlockTimeout: true,
+					useRpcCallSpecification: false,
+				},
 			});
 			expect(web3Eth.enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout).toBe(
 				true,
@@ -485,13 +489,42 @@ describe('defaults', () => {
 			// set by create new instance
 			eth2 = new Web3Eth({
 				config: {
-					enableExperimentalFeatures: { useSubscriptionWhenCheckingBlockTimeout: true },
+					enableExperimentalFeatures: {
+						useSubscriptionWhenCheckingBlockTimeout: true,
+						useRpcCallSpecification: false,
+					},
 				},
 			});
 			expect(eth2.enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout).toBe(
 				true,
 			);
 		});
+
+		it('enableExperimentalFeatures useRpcCallSpecification', () => {
+			// default
+			expect(web3Eth.enableExperimentalFeatures.useRpcCallSpecification).toBe(false);
+
+			// after set
+			web3Eth.setConfig({
+				enableExperimentalFeatures: {
+					useSubscriptionWhenCheckingBlockTimeout: false,
+					useRpcCallSpecification: true,
+				},
+			});
+			expect(web3Eth.enableExperimentalFeatures.useRpcCallSpecification).toBe(true);
+
+			// set by create new instance
+			eth2 = new Web3Eth({
+				config: {
+					enableExperimentalFeatures: {
+						useSubscriptionWhenCheckingBlockTimeout: false,
+						useRpcCallSpecification: true,
+					},
+				},
+			});
+			expect(eth2.enableExperimentalFeatures.useRpcCallSpecification).toBe(true);
+		});
+
 		it('should fallback to polling if provider support `on` but `newBlockHeaders` does not arrive in `blockHeaderTimeout` seconds', async () => {
 			const tempAcc2 = await createTempAccount();