+++ CHANGELOG.md
@@ -1136,3 +1136,5 @@ should use 4.0.1-alpha.0 for testing.
 -   Add `TransactionWithFromLocalWalletIndex`, `TransactionWithToLocalWalletIndex` and `TransactionWithFromAndToLocalWalletIndex` types (#5731)
 
 ## [Unreleased]
+
+-   Added rpc exception codes following eip-1474 as an experimental feature (if `useRpcCallSpecification` at `enableExperimentalFeatures` is `true`) (#5525)